OpenCL GPU vRAM headroom: make it a GUI global user pref

+++ src/common/opencl.h
@@ -184,7 +184,7 @@ typedef struct dt_opencl_device_t
 
   // Some devices are known to be unused by other apps so there is no need to test for available memory at all.
   // Also some devices might behave badly with the checking code, in this case we could enforce a headroom here.
-  int forced_headroom;
+  size_t forced_headroom;
 } dt_opencl_device_t;
 
 struct dt_bilateral_cl_global_t;
